nativetun: refactor: Split implementation into per-platform files
