+++ cmd/nativetun.go
@@ -1,0 +1,176 @@
+package cmd
+
+import (
+	"context"
+	"log"
+	"net"
+	"time"
+
+	"github.com/Diniboy1123/usque/api"
+	"github.com/Diniboy1123/usque/config"
+	"github.com/Diniboy1123/usque/internal"
+	"github.com/spf13/cobra"
+)
+
+type tunDevice struct {
+	name     string
+	mtu      int
+	iproute2 bool
+	ipv4     bool
+	ipv6     bool
+}
+
+var nativeTunCmd = &cobra.Command{
+	Use:   "nativetun",
+	Short: "Expose Warp as a native TUN device",
+	Long:  longDescription,
+	Run: func(cmd *cobra.Command, args []string) {
+		if !config.ConfigLoaded {
+			cmd.Println("Config not loaded. Please register first.")
+			return
+		}
+
+		sni, err := cmd.Flags().GetString("sni-address")
+		if err != nil {
+			cmd.Printf("Failed to get SNI address: %v\n", err)
+			return
+		}
+
+		privKey, err := config.AppConfig.GetEcPrivateKey()
+		if err != nil {
+			cmd.Printf("Failed to get private key: %v\n", err)
+			return
+		}
+		peerPubKey, err := config.AppConfig.GetEcEndpointPublicKey()
+		if err != nil {
+			cmd.Printf("Failed to get public key: %v\n", err)
+			return
+		}
+
+		cert, err := internal.GenerateCert(privKey, &privKey.PublicKey)
+		if err != nil {
+			cmd.Printf("Failed to generate cert: %v\n", err)
+			return
+		}
+
+		tlsConfig, err := api.PrepareTlsConfig(privKey, peerPubKey, cert, sni)
+		if err != nil {
+			cmd.Printf("Failed to prepare TLS config: %v\n", err)
+			return
+		}
+
+		keepalivePeriod, err := cmd.Flags().GetDuration("keepalive-period")
+		if err != nil {
+			cmd.Printf("Failed to get keepalive period: %v\n", err)
+			return
+		}
+		initialPacketSize, err := cmd.Flags().GetUint16("initial-packet-size")
+		if err != nil {
+			cmd.Printf("Failed to get initial packet size: %v\n", err)
+			return
+		}
+
+		connectPort, err := cmd.Flags().GetInt("connect-port")
+		if err != nil {
+			cmd.Printf("Failed to get connect port: %v\n", err)
+			return
+		}
+
+		var endpoint *net.UDPAddr
+		if ipv6, err := cmd.Flags().GetBool("ipv6"); err == nil && !ipv6 {
+			endpoint = &net.UDPAddr{
+				IP:   net.ParseIP(config.AppConfig.EndpointV4),
+				Port: connectPort,
+			}
+		} else {
+			endpoint = &net.UDPAddr{
+				IP:   net.ParseIP(config.AppConfig.EndpointV6),
+				Port: connectPort,
+			}
+		}
+
+		tunnelIPv4, err := cmd.Flags().GetBool("no-tunnel-ipv4")
+		if err != nil {
+			cmd.Printf("Failed to get no tunnel IPv4: %v\n", err)
+			return
+		}
+
+		tunnelIPv6, err := cmd.Flags().GetBool("no-tunnel-ipv6")
+		if err != nil {
+			cmd.Printf("Failed to get no tunnel IPv6: %v\n", err)
+			return
+		}
+
+		mtu, err := cmd.Flags().GetInt("mtu")
+		if err != nil {
+			cmd.Printf("Failed to get MTU: %v\n", err)
+			return
+		}
+		if mtu != 1280 {
+			log.Println("Warning: MTU is not the default 1280. This is not supported. Packet loss and other issues may occur.")
+		}
+
+		setIproute2, err := cmd.Flags().GetBool("no-iproute2")
+		if err != nil {
+			cmd.Printf("Failed to get no set address: %v\n", err)
+			return
+		}
+
+		reconnectDelay, err := cmd.Flags().GetDuration("reconnect-delay")
+		if err != nil {
+			cmd.Printf("Failed to get reconnect delay: %v\n", err)
+			return
+		}
+
+		interfaceName, err := cmd.Flags().GetString("interface-name")
+		if err != nil {
+			cmd.Printf("Failed to get interface name: %v\n", err)
+			return
+		}
+
+		if interfaceName != "" {
+			err = internal.CheckIfname(interfaceName)
+			if err != nil {
+				log.Printf("Invalid interface name: %v", err)
+				return
+			}
+		}
+
+		t := &tunDevice{
+			name:     interfaceName,
+			mtu:      mtu,
+			iproute2: !setIproute2,
+			ipv4:     !tunnelIPv4,
+			ipv6:     !tunnelIPv6,
+		}
+
+		dev, err := t.create()
+		if err != nil {
+			log.Println("Are you root/administrator? TUN device creation usually requires elevated privileges.")
+			log.Fatalf("Failed to create TUN device: %v", err)
+		}
+
+		log.Printf("Created TUN device: %s", t.name)
+
+		go api.MaintainTunnel(context.Background(), tlsConfig, keepalivePeriod, initialPacketSize, endpoint, dev, mtu, reconnectDelay)
+
+		log.Println("Tunnel established, you may now set up routing and DNS")
+
+		select {}
+	},
+}
+
+func init() {
+	nativeTunCmd.Flags().IntP("connect-port", "P", 443, "Used port for MASQUE connection")
+	nativeTunCmd.Flags().BoolP("ipv6", "6", false, "Use IPv6 for MASQUE connection")
+	nativeTunCmd.Flags().BoolP("no-tunnel-ipv4", "F", false, "Disable IPv4 inside the MASQUE tunnel")
+	nativeTunCmd.Flags().BoolP("no-tunnel-ipv6", "S", false, "Disable IPv6 inside the MASQUE tunnel")
+	nativeTunCmd.Flags().StringP("sni-address", "s", internal.ConnectSNI, "SNI address to use for MASQUE connection")
+	nativeTunCmd.Flags().DurationP("keepalive-period", "k", 30*time.Second, "Keepalive period for MASQUE connection")
+	nativeTunCmd.Flags().IntP("mtu", "m", 1280, "MTU for MASQUE connection")
+	nativeTunCmd.Flags().Uint16P("initial-packet-size", "i", 1242, "Initial packet size for MASQUE connection")
+	nativeTunCmd.Flags().BoolP("no-iproute2", "I", false, "Linux only: Do not set up IP addresses and do not set the link up")
+	nativeTunCmd.Flags().DurationP("reconnect-delay", "r", 1*time.Second, "Delay between reconnect attempts")
+	nativeTunCmd.Flags().StringP("interface-name", "n", "", "Custom inteface name for the TUN interface")
+	rootCmd.AddCommand(nativeTunCmd)
+}
