+++ cmd/nativetun_windows.go
@@ -3,191 +3,55 @@
 package cmd
 
 import (
-	"context"
-	"log"
-	"net"
-	"runtime"
-	"time"
+	"fmt"
 
 	"github.com/Diniboy1123/usque/api"
 	"github.com/Diniboy1123/usque/config"
 	"github.com/Diniboy1123/usque/internal"
-	"github.com/spf13/cobra"
 	"golang.zx2c4.com/wireguard/tun"
 )
 
-var nativeTunCmd = &cobra.Command{
-	Use:   "nativetun",
-	Short: "Expose Warp as a native TUN device",
-	Long: "Windows only: Expose Warp as a native TUN device. That accepts any IP traffic." +
-		" Requires wintun.dll and admins right.",
-	Run: func(cmd *cobra.Command, args []string) {
-		if !config.ConfigLoaded {
-			cmd.Println("Config not loaded. Please register first.")
-			return
-		}
+var longDescription = "Expose Warp as a native TUN device that accepts any IP traffic." +
+	" Requires wintun.dll and administrator rights."
 
-		if runtime.GOOS != "windows" {
-			cmd.Println("This command is only supported on Windows")
-		}
+func (t *tunDevice) create() (api.TunnelDevice, error) {
+	if t.name == "" {
+		t.name = "usque"
+	}
 
-		sni, err := cmd.Flags().GetString("sni-address")
+	dev, err := tun.CreateTUN(t.name, t.mtu)
+	if err != nil {
+		return nil, err
+	}
+
+	t.name, err = dev.Name()
+	if err != nil {
+		return nil, err
+	}
+
+	if t.ipv4 {
+		err = internal.SetIPv4Address(t.name, config.AppConfig.IPv4, "255.255.255.255")
 		if err != nil {
-			cmd.Printf("Failed to get SNI address: %v\n", err)
-			return
+			return nil, fmt.Errorf("failed to set IPv4 address: %v", err)
 		}
 
-		privKey, err := config.AppConfig.GetEcPrivateKey()
+		err = internal.SetIPv4MTU(t.name, t.mtu)
 		if err != nil {
-			cmd.Printf("Failed to get private key: %v\n", err)
-			return
+			return nil, fmt.Errorf("failed to set IPv4 MTU: %v", err)
 		}
-		peerPubKey, err := config.AppConfig.GetEcEndpointPublicKey()
+	}
+
+	if t.ipv6 {
+		err = internal.SetIPv6Address(t.name, config.AppConfig.IPv6, "128")
 		if err != nil {
-			cmd.Printf("Failed to get public key: %v\n", err)
-			return
+			return nil, fmt.Errorf("failed to set IPv6 address: %v", err)
 		}
 
-		cert, err := internal.GenerateCert(privKey, &privKey.PublicKey)
+		err = internal.SetIPv6MTU(t.name, t.mtu)
 		if err != nil {
-			cmd.Printf("Failed to generate cert: %v\n", err)
-			return
+			return nil, fmt.Errorf("failed to set IPv6 MTU: %v", err)
 		}
+	}
 
-		tlsConfig, err := api.PrepareTlsConfig(privKey, peerPubKey, cert, sni)
-		if err != nil {
-			cmd.Printf("Failed to prepare TLS config: %v\n", err)
-			return
-		}
-
-		keepalivePeriod, err := cmd.Flags().GetDuration("keepalive-period")
-		if err != nil {
-			cmd.Printf("Failed to get keepalive period: %v\n", err)
-			return
-		}
-		initialPacketSize, err := cmd.Flags().GetUint16("initial-packet-size")
-		if err != nil {
-			cmd.Printf("Failed to get initial packet size: %v\n", err)
-			return
-		}
-
-		connectPort, err := cmd.Flags().GetInt("connect-port")
-		if err != nil {
-			cmd.Printf("Failed to get connect port: %v\n", err)
-			return
-		}
-
-		var endpoint *net.UDPAddr
-		if ipv6, err := cmd.Flags().GetBool("ipv6"); err == nil && !ipv6 {
-			endpoint = &net.UDPAddr{
-				IP:   net.ParseIP(config.AppConfig.EndpointV4),
-				Port: connectPort,
-			}
-		} else {
-			endpoint = &net.UDPAddr{
-				IP:   net.ParseIP(config.AppConfig.EndpointV6),
-				Port: connectPort,
-			}
-		}
-
-		tunnelIPv4, err := cmd.Flags().GetBool("no-tunnel-ipv4")
-		if err != nil {
-			cmd.Printf("Failed to get no tunnel IPv4: %v\n", err)
-			return
-		}
-
-		tunnelIPv6, err := cmd.Flags().GetBool("no-tunnel-ipv6")
-		if err != nil {
-			cmd.Printf("Failed to get no tunnel IPv6: %v\n", err)
-			return
-		}
-
-		mtu, err := cmd.Flags().GetInt("mtu")
-		if err != nil {
-			cmd.Printf("Failed to get MTU: %v\n", err)
-			return
-		}
-		if mtu != 1280 {
-			log.Println("Warning: MTU is not the default 1280. This is not supported. Packet loss and other issues may occur.")
-		}
-
-		reconnectDelay, err := cmd.Flags().GetDuration("reconnect-delay")
-		if err != nil {
-			cmd.Printf("Failed to get reconnect delay: %v\n", err)
-			return
-		}
-
-		interfaceName, err := cmd.Flags().GetString("interface-name")
-		if err != nil {
-			cmd.Printf("Failed to get interface name: %v\n", err)
-			return
-		}
-
-		if interfaceName != "" {
-			err = internal.CheckIfname(interfaceName)
-			if err != nil {
-				log.Printf("Invalid interface name: %v", err)
-				return
-			}
-		} else {
-			interfaceName = "usque"
-		}
-
-		dev, err := tun.CreateTUN(interfaceName, mtu)
-		if err != nil {
-			log.Println("Are you administrator? TUN device creation usually requires elevated privileges.")
-			log.Fatalf("failed to create TUN device: %v", err)
-		}
-
-		devName, err := dev.Name()
-		if err != nil {
-			log.Fatalf("failed to get TUN device name: %v", err)
-		}
-
-		log.Printf("Created TUN device: %s", devName)
-
-		if !tunnelIPv4 {
-			err = internal.SetIPv4Address(devName, config.AppConfig.IPv4, "255.255.255.255")
-			if err != nil {
-				log.Fatalf("failed to set IPv4 address: %v", err)
-			}
-
-			err = internal.SetIPv4MTU(devName, mtu)
-			if err != nil {
-				log.Fatalf("failed to set IPv4 MTU: %v", err)
-			}
-		}
-
-		if !tunnelIPv6 {
-			err = internal.SetIPv6Address(devName, config.AppConfig.IPv6, "128")
-			if err != nil {
-				log.Fatalf("failed to set IPv6 address: %v", err)
-			}
-
-			err = internal.SetIPv6MTU(devName, mtu)
-			if err != nil {
-				log.Fatalf("failed to set IPv6 MTU: %v", err)
-			}
-		}
-
-		go api.MaintainTunnel(context.Background(), tlsConfig, keepalivePeriod, initialPacketSize, endpoint, api.NewNetstackAdapter(dev), mtu, reconnectDelay)
-
-		log.Println("Tunnel established, you may now set up routing and DNS")
-
-		select {}
-	},
-}
-
-func init() {
-	nativeTunCmd.Flags().IntP("connect-port", "P", 443, "Used port for MASQUE connection")
-	nativeTunCmd.Flags().BoolP("ipv6", "6", false, "Use IPv6 for MASQUE connection")
-	nativeTunCmd.Flags().BoolP("no-tunnel-ipv4", "F", false, "Disable IPv4 inside the MASQUE tunnel")
-	nativeTunCmd.Flags().BoolP("no-tunnel-ipv6", "S", false, "Disable IPv6 inside the MASQUE tunnel")
-	nativeTunCmd.Flags().StringP("sni-address", "s", internal.ConnectSNI, "SNI address to use for MASQUE connection")
-	nativeTunCmd.Flags().DurationP("keepalive-period", "k", 30*time.Second, "Keepalive period for MASQUE connection")
-	nativeTunCmd.Flags().IntP("mtu", "m", 1280, "MTU for MASQUE connection")
-	nativeTunCmd.Flags().Uint16P("initial-packet-size", "i", 1242, "Initial packet size for MASQUE connection")
-	nativeTunCmd.Flags().DurationP("reconnect-delay", "r", 1*time.Second, "Delay between reconnect attempts")
-	nativeTunCmd.Flags().StringP("interface-name", "n", "", "Custom inteface name for the TUN interface")
-	rootCmd.AddCommand(nativeTunCmd)
+	return api.NewNetstackAdapter(dev), nil
 }
